fix(devops): Fix local bitcoin deployment (#59)
# Motivation
`dfx deploy` is breaking erratically. `dfx deploy bitcoin` is breaking
consistently. This is because it is using a remote canister id for
`local`, instead of a `specified_id`.

# Changes
- Remove the bitcoin remote canister id for `local`.
- Add a specified canister id for bitcoin.
- Add a default init arg for bitcoin.

# Tests
`dfx deploy bitcoin` now works, even ina completely clean repo. It is
yet to be seen whether this resolves the erratic `dfx deploy` breakages.