feat(signer): btc_caller_send endpoint (#108)
# Motivation

Add endpoint to create, sign and send bitcoin tokens.

# Changes

* New api types for the new endpoint `SendBtcRequest`, `SendBtcResponse`
and `SendBtcError`.
* New endpoint `btc_caller_send` with new request, response and error
types. It uses different helpers to create a transaction, then sign it
and finally send it.
* New `fee_utils.rs` module with public function `calculate_fee`. This
module is necessary only if we want to have a smart choice for the fee
when not present. We could just set it as mandatory and avoid all this
code.
* New `ecds_api.rs` module with `get_ecdsa_signature` and
`ecdsa_pubkey_of`. It interacts with the ECDSA api from Mgt canister in
an agnostic way from ETH or BTC.
* New `tx_utils.rs` module with `build_p2wpkh_transaction` and
`btc_sign_transaction`.
* Add methods to `bitcoin_api.rs` module `get_fee_per_byte` and
`send_transaction`.
* Expose `transform_network` from `bitcoin_utils.rs`.

## Code Review Order Suggestion

1. Start with the new endpoint and the logic it follows in
`src/signer/canister/src/lib.rs`.
2. You can take a look at the types used in the interface in
`src/signer/api/src/types.rs`.
3. Review the function that creates the transaction
[build_p2wpkh_transaction](https://github.com/dfinity/chain-fusion-signer/pull/108/files#diff-55d13bf9e7a4d9ceb936027707f4ec65435c10dff3b2e317690f91b8ef64205cR58).
Follow the helpers used in it.
4. Review the function that signs the transaction
[btc_sign_transaction](https://github.com/dfinity/chain-fusion-signer/pull/108/files#diff-55d13bf9e7a4d9ceb936027707f4ec65435c10dff3b2e317690f91b8ef64205cR140).
Follow the hepers used in it.

# Tests

Waiting for pocket-ic to support bitcoin to add integration tests.