feat(signer): Better error reporting in btc_caller_send (#121)
# Motivation

Clients should be able to differentiate errors when sending bitcoins and
act differently depending on the error.

In this PR, I introduce one more enum in the error response for
`btc_caller_send`: `BuildP2wpkhError`.

# Changes

* New enum `BuildP2wpkhTxError` for error reporting.
* Update signer candid file.
* `build_p2wpkh_transaction` returns a `BuildP2wpkhTxError` instead of a
String if error.

# Tests

Unfortunately we can't add tests for this yet. We need to upgrade
pocket-ic when this supports bitcoin.