refactor(devops): Move integration test (#126)
# Motivation
The integration test will soon require most of the canisters in
`dfx.json`. A simple and efficient way of doing this is to run the
integration test just after the `dfx deploy` test, which fetches or
builds all the relevant Wasms anyway.

# Changes
- Move the integration test to just after dfx deploy.

Note: We can simplify the integration test script (e.g. the bitcoin
canister Wasm is already present) but that involves some test changes as
well, so is better done in a separate PR.

# Tests
Existing CI suffices.

Note: I have configured the integration test to be no longer required
for merge.