test(signer): Extend the pic canister trait (#155)
# Motivation
The pocket_ic canister tests will need to manage multiple canisters, not
just one. The utilities used to do this, to date, only support one
canister.

# Changes
- Move the PicCanister trait into a separate file and add more methods
to it.
- Create a more general version of the canister builder, suitable for
use with multiple canisters.

# Tests
This is used in tests in the `papi` branch, but I am merging this
separately to keep the PRs small and reviewable.