refactor(devops): Compress all Wasms (#144)
# Motivation
Currently, when deploying to local testnets, not all Wasms are
compressed. This makes paths to the Wasm builds inconsistent - some end
in `.wasm` and others in `.wasm.gz`. This is an issue if we import the
pocket-ic papi tests that always assume `.wasm.gz`

# Changes
- Compress all Wasms

# Tests
Deploying to localhost, we see that all Wasms are compressed:
```
$ dfx deploy
$ find .dfx/local/canisters/ | grep wasm | grep -v wasm.gz
```