refactor(signer): Rename update->update_one (#163)
# Motivation
The `update()` utility used in `pocket_ic` tests only supports canister
methods with a single argument and a single return value. This is a
common case but not the only one.

# Changes
- Rename `update(..)` to `update_one(..)`, to match functions such as
`encode_one()` and `decode_one()`. This makes space for `update()` to
cover general update calls.

# Tests
Existing tests should suffice.