feat(signer): Add pocket-ic bindings for the cycles ledger (#162)
# Motivation
In tests, we will need to call the cycles ledger.

# Changes
- Add support for zero or one arguments to the handlebars rust binding
template. Currently only one argument is supported and some ledger
functions take no arguments.
- Add bindings to the cycles ledger

# Tests
There are existing tests that the bindings are up to date and that the
rust code compiles. But the code is not used yet in any way; it will be
used in tests.