feat(devops): Create beta network (#267)
# Motivation
It is useful to have a test canister running on the same subnet as
production, to get accurate pricing and behaviour metrics.

The staging canister operates on a different subnet, and is used
extensively by projects other than the chain fusion signer project
itself, making it hard to get precise measurements from that canister.

# Changes
- Define a new beta environment
- Create a new signer canister, on the same subnet as production.

# Tests
## Claim: The beta canister is on the same subnet as production.
This can be checked by comparing the subnets on the dashboard.

Beta:
https://dashboard.internetcomputer.org/canister/uhxml-hyaaa-aaaar-qaoxq-cai
Prod:
https://dashboard.internetcomputer.org/canister/grghe-syaaa-aaaar-qabyq-cai

## Claim: The new canister is supplied with cycles
The new canister is configured in cycleops under our team account.

## Claim: The team has control over the canister
The canister has Orbit as a controller.