fix(devops): ENV -> DFX_NETWORK (#271)
# Motivation
We now have a beta network for the chain fusion signer, but `dfx build
staging --network beta` fails.

The deploy code works for `local`, `staging` and `ic` but not for other
networks.

There is a match clause that deals with `ic`, `staging` and everything
else. The branch for "everything else" gets the network from "ENV",
defaulting to `local` if `ENV` is not defined. So it worked for local
but not for `beta`. The `ENV` should be `DFX_NETWORK`.

# Changes
- Correct `ENV` to `DFX_NETWORK`.


# Tests

`dfx build --network beta` now works.